Create sql code
There are 3 tables which are the Guests, Rooms, and Reservations that represent the hotel Database.
The tables are populated with data and are ready to reverse engineered in mysql workbench. 
These tables are all connected with a many-to-many relationship with each other.
A room can part of many reservations and vice versa. 
A guest can also have multiple reservations and vice versa.